pyproject.toml
[project]
name = "kindlenotes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.129.0",
    "genanki>=0.13.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
